CB-13347 Increase sizelimit and timelimit for retrieving users and groups from FreeIPA
User sync retrieves all the users and groups from IPA for comparison against UMS. We need to increase the nsslapd-sizelimit configuration in preparation for supporting more than 2000 users or groups. If we don't make this change, then user sync will think that it needs to add users and groups that already exist in IPA.

If we don't increases this, then the IPA user-find and group-find will truncate the results to 2000 entries. This sizelimit is always enforced even if we set the sizelimit=0 in the request.

The new limit is 11000 to ensure the customer 10000 users+other stuff can fit into the limit.
GZIP compression support is turned on for responses, as without it for 10600 user we have seen 11.7MB responses. With GZIP support it's reduced to 0.5MB

Tested with 10000+ users:
- DL+DH create/termination with and without cluster proxy
- with single and multi node FreeIPA
- `ipa user-find --sizelimit 0 --timelimit 0` is finished under 30 sec, closer to 20 sec